+++ src/pages/profile-other/profile.styled.tsx
@@ -8,6 +8,20 @@ export const Wrapper = styled.div`
   width: 100%;
   height: auto;
   background-color: #f5f5f5;
+
+  animation: fadein 2s;
+  -moz-animation: fadein 2s; /* Firefox */
+  -webkit-animation: fadein 2s; /* Safari and Chrome */
+  -o-animation: fadein 2s; /* Opera */
+
+  @keyframes fadein {
+    from {
+      opacity: 0;
+    }
+    to {
+      opacity: 1;
+    }
+  }
 `;
 
 export const ContentWrapper = styled.div`
