+++ src/pages/trade-post/index.tsx
@@ -1,4 +1,6 @@
-import { useNavigate } from 'react-router-dom';
+import { useEffect, useState } from 'react';
+import { useNavigate, useParams } from 'react-router-dom';
+import axios from 'axios';
 
 import Gnb from '../../components/gnb';
 import Spinner from '../../components/spinner';
@@ -7,13 +9,42 @@ import OtherTrades from './container/other-trades';
 import ContentFooter from '../../components/content-footer';
 
 import { useAuth } from '../../hooks/useAuth';
+import { loadItem } from '../../utils/storage';
+import { useAppDispatch } from '../../store/hooks';
 import { normalToast } from '../../utils/basic-toast-modal';
+import { getTradePost } from '../../store/slices/tradePostSlice';
 
 import * as S from './trade-post.styled';
 
 const TradePostPage = () => {
   const navigate = useNavigate();
+  const dispatch = useAppDispatch();
+  const postId = Number(useParams().id);
+  const [dataLoading, setDataLoading] = useState(true);
   const { sessionLoading, isAuthed } = useAuth();
+  const accessToken = loadItem('accessToken');
+
+  useEffect(() => {
+    if (accessToken && postId) {
+      dispatch(getTradePost({ accessToken, postId }))
+        .unwrap()
+        .then(() => {
+          setDataLoading(false);
+        })
+        .catch(err => {
+          if (axios.isAxiosError(err)) {
+            if (err.response?.status === 404) {
+              normalToast(err.response?.data.error);
+            } else {
+              normalToast('요청을 수행할 수 없습니다.');
+              navigate(-1);
+            }
+          }
+        });
+    }
+  }, [accessToken, postId]);
+
+  window.scrollTo({ top: 0, behavior: 'smooth' });
 
   if (!sessionLoading && !isAuthed) {
     navigate('/login');
@@ -24,7 +55,9 @@ const TradePostPage = () => {
     return <Spinner />;
   }
 
-  window.scrollTo({ top: 0, behavior: 'smooth' });
+  if (dataLoading) {
+    return <Spinner />;
+  }
 
   return (
     <S.Wrapper>
