+++ src/pages/trade-post/trade-post.styled.tsx
@@ -10,6 +10,20 @@ export const Wrapper = styled.div`
   height: auto;
   min-width: 360px;
   background-color: #f5f5f5;
+
+  animation: fadein 0.5s;
+  -moz-animation: fadein 0.5s; /* Firefox */
+  -webkit-animation: fadein 0.5s; /* Safari and Chrome */
+  -o-animation: fadein 0.5s; /* Opera */
+
+  @keyframes fadein {
+    from {
+      opacity: 0;
+    }
+    to {
+      opacity: 1;
+    }
+  }
 `;
 
 export const ContentWrapper = styled.div`
